conanfile.py: make _export_local_recipes more robust
By using 'exports' instead of 'exports_sources',
which copies the recipes to the recipe dir instead of exported sources directory.

This should fix any potential issues with using the bundled recipes when using the cupoch recipe from a Conan repo, in which case only the recipe contents are downloaded and 'exports_sources' content is unavailable.